requirements.txt
# PySpark and core dependencies
pyspark>=3.4.0

# Data processing
pandas>=2.0.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# API fetching
requests>=2.31.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
chispa>=0.9.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
